+++ crates/cli/tests/sample-scripts/plugin-exports.js
@@ -1,0 +1,5 @@
+export function fn() {
+    if (plugin !== true) {
+        throw new Error("Not using the test_plugin");
+    }
+}