+++ crates/codegen/src/wit.rs
@@ -1,6 +1,6 @@
 use std::path::{Path, PathBuf};
 
-use anyhow::{bail, Result};
+use anyhow::{Result, bail};
 
 use wit_parser::{Resolve, WorldItem};