+++ crates/codegen/tests/integration_test.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use anyhow::Result;
-use javy_codegen::{Generator, LinkingKind, Plugin, WitOptions, JS};
+use javy_codegen::{Generator, JS, LinkingKind, Plugin, WitOptions};
 
 #[test]
 fn test_empty() -> Result<()> {
@@ -57,7 +57,7 @@ fn default_plugin() -> Result<Plugin> {
             .join("..")
             .join("..")
             .join("target")
-            .join("wasm32-wasip2")
+            .join("wasm32-wasip1")
             .join("release")
             .join("plugin_wizened.wasm"),
     )