+++ crates/plugin/src/lib.rs
@@ -1,41 +1,13 @@
 use std::io::{self, Read};
-use std::process;
 
 use javy_plugin_api::javy::Runtime;
-use javy_plugin_api::{import_namespace, Config};
+use javy_plugin_api::{Config, import_namespace};
 
 use crate::shared_config::SharedConfig;
 
 mod shared_config;
 
-wit_bindgen::generate!({ world: "javy-default-plugin", generate_all });
-
 import_namespace!("javy-default-plugin-v2");
-
-struct Component;
-
-impl Guest for Component {
-    fn config_schema() -> Vec<u8> {
-        shared_config::config_schema()
-    }
-
-    fn invoke(bytecode: Vec<u8>, function: Option<String>) {
-        javy_plugin_api::invoke(&bytecode, function.as_deref()).unwrap_or_else(|e| {
-            eprintln!("{e}");
-            process::abort();
-        })
-    }
-
-    fn compile_src(src: Vec<u8>) -> Result<Vec<u8>, String> {
-        javy_plugin_api::compile_src(&src).map_err(|e| e.to_string())
-    }
-
-    fn initialize_runtime() {
-        javy_plugin_api::initialize_runtime(config, modify_runtime).unwrap()
-    }
-}
-
-export!(Component);
 
 fn config() -> Config {
     // Read shared config JSON in from stdin.
@@ -63,4 +35,9 @@ fn config() -> Config {
 
 fn modify_runtime(runtime: Runtime) -> Runtime {
     runtime
+}
+
+#[unsafe(export_name = "initialize-runtime")]
+fn initialize_runtime() {
+    javy_plugin_api::initialize_runtime(config, modify_runtime).unwrap();
 }